// src/features/characters/scouter/bosscut-data.generated.ts
export interface BossCutEntry {
  boss: string;
  name: string;
  difficulty: string;
  level: number;
  guard: number;
  arcaneForce: number | null;
  authenticForce: number | null;
  partyLimit: number;
  bossCut: number | null;
  partyBossCut: number | null;
  easyRate: number | null;
  challenger: number | null;
  renewalDate: string;
  renewalDetail: string;
}